% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sub.R
\name{get_sub_sections}
\alias{get_sub_sections}
\title{Get sub sections}
\usage{
get_sub_sections(suffix)
}
\arguments{
\item{suffix}{A character string. The sections suffix. Needs to start and end
with a slash.}
}
\value{
A tibble with two columns: 'name' and 'suffix'. The former indicates
the section's name, the latter its link's suffix.
}
\description{
Returns the sub sections' names and suffixes
}
\examples{
get_sub_sections(suffix = "/Forum-4-0/")

}
