% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_threads.R
\name{get_threads}
\alias{get_threads}
\title{Get threads
Returns the thread links}
\usage{
get_threads(
  suffix,
  thread_start_date = "2000-01-01",
  latest_entry = "2000-01-01"
)
}
\arguments{
\item{suffix}{A character string. The section's suffix.}

\item{thread_start_date}{A character string of a date in the format
"YYYY-MM-DD". Links to threads which were started before this date will not
be returned. Defaults to "2000-01-01".}

\item{latest_entry}{A character string of a date in the format "YYYY-MM-DD".
Only links to threads whose latest entry was posted on or after this very
date are returned.}
}
\value{
A tibble with three columns. \code{link} contains the thread's link,
\code{start_date} the date it was created on, and \code{latest_entry} the
date when the latest posting was made within the thread.
}
\description{
Get threads
Returns the thread links
}
\examples{
get_threads(suffix = "/Forum-19-89/", thread_start_date = "2020-04-01", latest_entry = "2020-05-01")

}
