% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_thread.R
\name{scrape_thread}
\alias{scrape_thread}
\title{Scrape thread
Returns the threas' content}
\usage{
scrape_thread(thread_link)
}
\arguments{
\item{thread_link}{A character string. The thread's link.}
}
\value{
A tibble with a bunch of columns. \code{link} contains the thread's link,
\code{start_date} the date it was created on, and \code{latest_entry} the
date when the latest posting was made within the thread.
}
\description{
Scrape thread
Returns the threas' content
}
\examples{
get_threads(suffix = "/Forum-19-89/", thread_start_date = "2020-04-01", latest_entry = "2020-05-01")

}
