% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_thread.R
\name{scrape_thread}
\alias{scrape_thread}
\title{Scrape thread
Returns the threads' content}
\usage{
scrape_thread(thread_link)
}
\arguments{
\item{thread_link}{A character string. The thread's link.}
}
\value{
A tibble with a bunch of columns. \code{thread} contains the url
where you can find the thread containing the posting, \code{date} the date
the posting was created on, \code{time} the time it was created at,
\code{content} its textual content, and \code{quote_ind} indicates whether
it contains quoted content or not. Unfortunately, it is nearly impossible to
remove the quotes in a reasonable manner.
}
\description{
Scrape thread
Returns the threads' content
}
\examples{
scrape_thread(thread_link)

}
